pyproject.toml
[project]
name = "task-service"
version = "0.1.0"
description = ""
authors = [
    {name = "TheAppleKingy",email = "pridanov.ali.job@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "fastapi (>=0.116.0,<0.117.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "pytest (>=8.4.1,<9.0.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "pytest-asyncio (>=1.0.0,<2.0.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "bcrypt (>=4.3.0,<5.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "freezegun (>=1.5.2,<2.0.0)",
    "alembic (>=1.16.2,<2.0.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "celery (>=5.5.3,<6.0.0)",
    "redis (>=6.2.0,<7.0.0)",
    "flower (>=2.0.1,<3.0.0)",
    "itsdangerous (>=2.2.0,<3.0.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "fastapi-cachetools (>=0.1.0,<0.2.0)",
    "pytest-mock (>=3.14.1,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
